LayoutTests/css3/flexbox/width-change-and-relayout-children.html

<!DOCTYPE html>
<html>
<head>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexitem {
    width: 100%;
    height: 3em;
    min-height: 3em;
}
.child {
    background-color: salmon;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
</head>
<body>
<div id=log></div>
<div id="container" style="width: 100px" data-expected-width="200">
<div class="flexbox column" data-expected-width="200">
    <div class="flexitem" data-expected-width="200">
        <div class="child" data-expected-width="200">This div should be 200px wide.</div>
    </div>
</div>
</div>
<script>
var container = document.getElementById('container');
container.offsetWidth;
container.style.width = "200px";
window.onload = function()
{
    checkLayout("#container");
};
</script>
</body>
</html>
